chore: rename animals service and tests to animal
