+++ test/services/animal.spec.js
